services/clip2video-vtr/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic-settings==2.0.3


# Models
gdown
torch==1.7.1
torchvision==0.8.2
ftfy==6.1.1
regex
tqdm==4.66.1
pandas==2.0.1
opencv-python

# Vector Database
faiss-cpu

# Project settings, for development only
# pre-commit